Core PROGMEM definitions provided by newlib in sys/pgmspace.h
Move corresponding definitions into libc for use with old toolchain
