+++ Sming/Arch/Esp8266/Components/libc/component.mk
@@ -7,6 +7,7 @@ COMPONENT_SRCFILES		+= libc_replacements.c
 LIBDIRS					+= $(COMPONENT_PATH)/newlib
 EXTRA_LIBS				+= m c gcc
 else
+COMPONENT_SRCFILES		+= pgmspace.c
 LIBDIRS					+= $(COMPONENT_PATH)/lib
 EXTRA_LIBS				+= microc microgcc setjmp
 endif
