+++ Sming/Wiring/FakePgmSpace.h
@@ -10,9 +10,11 @@
 
 #pragma once
 
+#include <esp_attr.h>
+#include <sys/pgmspace.h>
+
 #include "m_printf.h"
 #include "c_types.h"
-#include <esp_attr.h>
 
 #ifdef __cplusplus
 extern "C"
@@ -21,20 +23,6 @@ extern "C"
 
 // Simple check to determine if a pointer refers to flash memory
 #define isFlashPtr(ptr) (uint32_t(ptr) >= 0x40200000)
-
-#define PGM_P  const char *
-
-#define PRIPSTR "%s"
-
-typedef void prog_void;
-typedef char prog_char;
-typedef unsigned char prog_uchar;
-typedef int8_t prog_int8_t;
-typedef uint8_t prog_uint8_t;
-typedef int16_t prog_int16_t;
-typedef uint16_t prog_uint16_t;
-typedef int32_t prog_int32_t;
-typedef uint32_t prog_uint32_t;
 
 /** @brief determines if the given value is aligned to a word (4-byte) boundary */
 #define IS_ALIGNED(_x) (((uint32_t)(_x)&3) == 0)
@@ -55,76 +43,8 @@ typedef uint32_t prog_uint32_t;
 #endif
 
 
-#ifdef ICACHE_FLASH
-
-#ifndef PROGMEM
+#undef PROGMEM
 #define PROGMEM __attribute__((aligned(4))) ICACHE_FLASH_ATTR
-#endif
-
-// flash memory must be read using 32 bit aligned addresses else a processor exception will be triggered
-// order within the 32 bit values are
-// --------------
-// b3, b2, b1, b0
-//     w1,     w0
-
-#define pgm_read_with_offset(addr, res)                                                                                \
-	__asm__("extui    %0, %1, 0, 2\n" /* Extract offset within word (in bytes) */                                      \
-			"sub      %1, %1, %0\n"   /* Subtract offset from addr, yielding an aligned address */                     \
-			"l32i.n   %1, %1, 0x0\n"  /* Load word from aligned address */                                             \
-			"slli     %0, %0, 3\n"	/* Mulitiply offset by 8, yielding an offset in bits */                          \
-			"ssr      %0\n"			  /* Prepare to shift by offset (in bits) */                                       \
-			"srl      %0, %1\n"		  /* Shift right; now the requested byte is the first one */                       \
-			: "=r"(res), "=r"(addr)                                                                                    \
-			: "1"(addr)                                                                                                \
-			:);
-
-static inline uint8_t pgm_read_byte_inlined(const void* addr)
-{
-	register uint32_t res;
-	pgm_read_with_offset(addr, res);
-	return (uint8_t)res; /* This masks the lower byte from the returned word */
-}
-
-/* Although this says "word", it's actually 16 bit, i.e. half word on Xtensa */
-static inline uint16_t pgm_read_word_inlined(const void* addr)
-{
-	register uint32_t res;
-	pgm_read_with_offset(addr, res);
-	return (uint16_t)res; /* This masks the lower half-word from the returned word */
-}
-
-// Make sure, that libraries checking existence of this macro are not failing
-#define pgm_read_byte(addr) pgm_read_byte_inlined(addr)
-#define pgm_read_word(addr) pgm_read_word_inlined(addr)
-
-// No translation necessary (provided address is aligned)
-#define pgm_read_dword(addr) (*(const unsigned long*)(addr))
-#define pgm_read_float(addr) (*(const float*)(addr))
-
-void *memcpy_P(void *dest, const void *src_P, size_t length);
-int memcmp_P(const void *a1, const void *b1, size_t len);
-size_t strlen_P(const char * src_P);
-char *strcpy_P(char * dest, const char * src_P);
-char *strncpy_P(char * dest, const char * src_P, size_t size);
-int strcmp_P(const char *str1, const char *str2_P);
-int strncmp_P(const char *str1, const char *str2_P, const size_t size);
-int strcasecmp_P(const char* str1, const char* str2_P);
-char* strcat_P(char* dest, const char* src_P);
-char *strstr_P(char *haystack, const char *needle_P);
-
-#define sprintf_P(s, f_P, ...)                                                                                         \
-	(__extension__({                                                                                                   \
-		int len_P = strlen_P(f_P);                                                                                     \
-		int __result = 0;                                                                                              \
-		char* __localF = (char*)malloc(len_P + 1);                                                                     \
-		if(__localF) {                                                                                                 \
-			strcpy_P(__localF, f_P);                                                                                   \
-			__localF[len_P] = '\0';                                                                                    \
-		}                                                                                                              \
-		__result = m_snprintf(s, len_P, __localF, ##__VA_ARGS__);                                                      \
-		free(__localF);                                                                                                \
-		__result;                                                                                                      \
-	}))
 
 #define printf_P_heap(f_P, ...)                                                                                        \
 	(__extension__({                                                                                                   \
@@ -145,39 +65,6 @@ char *strstr_P(char *haystack, const char *needle_P);
 
 #define printf_P printf_P_stack
 
-#else /* ICACHE_FLASH */
-
-#define PROGMEM __attribute__((aligned(4)))
-
-#define pgm_read_byte(addr) (*(const unsigned char *)(addr))
-#define pgm_read_word(addr) (*(const unsigned short *)(addr))
-#define pgm_read_dword(addr) (*(const unsigned long *)(addr))
-#define pgm_read_float(addr) (*(const float *)(addr))
-
-#define memcpy_P(dest, src, num) memcpy((dest), (src), (num))
-#define memcmp_P(a1, b1, len) memcmp(a1, b1, len)
-#define strlen_P(a) strlen((a))
-#define strcpy_P(dest, src) strcpy((dest), (src))
-#define strncpy_P(dest, src, size) strncpy((dest), (src), (size))
-#define strcmp_P(a, b) strcmp((a), (b))
-#define strncmp_P(str1, str2_P, size) strncmp(str1, str2_P, size)
-#define strcasecmp_P(a, b) strcasecmp((a), (b))
-#define strcat_P(dest, src) strcat((dest), (src))
-#define strstr_P(a, b) strstr((a), (b))
-#define sprintf_P(s, f, ...) m_snprintf(s, 1024, f, ##__VA_ARGS__)
-#define printf_P(f, ...) m_printf((f), ##__VA_ARGS__)
-
-#endif /* ICACHE_FLASH */
-
-/*
- * Define and use a flash string inline
- */
-#define PSTR(str)                                                                                                     \
-	(__extension__({                                                                                                   \
-		DEFINE_PSTR_LOCAL(__c, str);                                                                                  \
-		&__c[0];                                                                                                       \
-	}))
-
 /*
  * Declare and use a flash string inline.
  * Returns a pointer to a stack-allocated buffer of the precise size required.
@@ -189,15 +76,6 @@ char *strstr_P(char *haystack, const char *needle_P);
 		_buf;                                                                                                          \
 	}))
 
-
-#define pgm_read_byte_near(addr) pgm_read_byte(addr)
-#define pgm_read_word_near(addr) pgm_read_word(addr)
-#define pgm_read_dword_near(addr) pgm_read_dword(addr)
-#define pgm_read_float_near(addr) pgm_read_float(addr)
-#define pgm_read_byte_far(addr) pgm_read_byte(addr)
-#define pgm_read_word_far(addr) pgm_read_word(addr)
-#define pgm_read_dword_far(addr) pgm_read_dword(addr)
-#define pgm_read_float_far(addr) pgm_read_float(addr)
 
 void* memcpy_aligned(void* dst, const void* src, unsigned len);
 int memcmp_aligned(const void* ptr1, const void* ptr2, unsigned len);
